@@ -14,17 +14,23 @@
         "tanner Campbell <tcamp022@gmail.com>"
     ],
     "require": {
-        "php": "^8.1",
-        "illuminate/support": "9.*|10.*",
-        "illuminate/database": "^10.0.0"
+        "php": "^8.2",
+        "illuminate/support": "^11.33.0|^12.0",
+        "illuminate/database": "^11.33.0|^12.0",
+        "illuminate/console": "^11.33.0|^12.0"
     },
     "require-dev": {
-        "spatie/laravel-ray": "^1.17",
-        "mockery/mockery": "^1.4.4",
-        "orchestra/testbench": "^7.0|^8.0",
-        "phpunit/phpunit": "^9.5|^10.0",
-        "laravel/pint": "^1.3",
-        "nunomaduro/larastan": "^2.2"
+        "consolidation/robo": "^5.1.0",
+        "larastan/larastan": "^3.0.2",
+        "laravel/pint": "^1.18.3",
+        "orchestra/testbench": "^9.6.1|^10.0",
+        "phpstan/phpstan-deprecation-rules": "^2.0.1",
+        "phpunit/phpunit": "^11.4.4",
+        "totten/lurkerlite": "^1.3"
+    },
+    "conflict": {
+        "laravel/framework": "<11.33.0",
+        "nesbot/carbon": "<3.5.0"
     },
     "autoload": {
         "psr-4": {
@@ -33,7 +39,10 @@
     },
     "autoload-dev": {
         "psr-4": {
-            "FumeApp\\FastPModelTyperaginate\\Tests\\": "tests/"
+            "Tests\\": "test/Tests/",
+            "App\\": "test/laravel-skeleton/app/",
+            "Database\\Factories\\": "test/laravel-skeleton/database/factories/",
+            "Database\\Seeders\\": "test/laravel-skeleton/database/seeders/"
         }
     },
     "extra": {
@@ -45,22 +54,38 @@
     },
     "scripts": {
         "pint": [
-            "vendor/bin/pint"
+            "@php vendor/bin/pint --ansi"
         ],
         "test": [
-            "./vendor/bin/testbench package:test --no-coverage"
+            "@php vendor/bin/phpunit --colors --display-errors --testdox"
+        ],
+        "test-watch": [
+            "@php vendor/bin/robo watch 'src, test/Tests' 'clear && composer test'"
         ],
         "test-coverage": [
-            "XDEBUG_MODE=coverage ./vendor/bin/phpunit --colors=always --testdox --coverage-text"
+            "@php XDEBUG_MODE=coverage ./vendor/bin/phpunit --colors=always --testdox --coverage-text"
         ],
         "stan": [
-            "vendor/bin/phpstan analyse --xdebug"
-        ]
+            "@php vendor/bin/phpstan analyse --verbose --ansi"
+        ],
+        "post-autoload-dump": [
+            "@clear",
+            "@prepare"
+        ],
+        "clear": "@php vendor/bin/testbench package:purge-skeleton --ansi",
+        "prepare": "@php vendor/bin/testbench package:discover --ansi"
     },
     "scripts-descriptions": {
         "pint": "Run the Pint Linter and Fixer.",
         "test": "Run the PHPUnit tests.",
         "test-coverage": "Run the PHPUnit tests with code coverage.",
-        "stan": "Run the PHPStan analyser."
-    }
+        "stan": "Run PHPStan analyzer."
+    },
+    "config": {
+        "optimize-autoloader": true,
+        "preferred-install": "dist",
+        "sort-packages": true
+    },
+    "minimum-stability": "stable",
+    "prefer-stable": true
 }
