@@ -5,6 +5,7 @@ namespace FumeApp\ModelTyper\Actions;
 use Exception;
 use Illuminate\Support\Str;
 use ReflectionClass;
+use ReflectionException;
 use ReflectionMethod;
 
 class DetermineAccessorType
@@ -13,30 +14,25 @@ class DetermineAccessorType
      * Determine the type of accessor.
      *
      * @see https://laravel.com/docs/9.x/eloquent-mutators#defining-an-accessor
-     * @see https://laravel.com/docs/8.x/eloquent-mutators#defining-an-accessor
      *
-     * @param  ReflectionClass  $reflectionModel
-     * @param  string  $mutator
-     * @return ReflectionMethod
+     * @param  \ReflectionClass<\Illuminate\Database\Eloquent\Model>  $reflectionModel
      *
      * @throws Exception
      */
     public function __invoke(ReflectionClass $reflectionModel, string $mutator): ReflectionMethod
     {
+        $mutator = Str::studly($mutator);
+
         // Try traditional
         try {
-            $accessor = 'get' . Str::studly($mutator) . 'Attribute';
-
-            return $reflectionModel->getMethod($accessor);
-        } catch (Exception $e) {
+            return $reflectionModel->getMethod('get' . $mutator . 'Attribute');
+        } catch (ReflectionException $e) {
         }
 
         // Try new
         try {
-            $method = Str::studly($mutator);
-
-            return $reflectionModel->getMethod($method);
-        } catch (Exception $e) {
+            return $reflectionModel->getMethod($mutator);
+        } catch (ReflectionException $e) {
         }
 
         throw new Exception('Accessor method for ' . $mutator . ' on model ' . $reflectionModel->getName() . ' does not exist');
