@@ -5,12 +5,19 @@ namespace FumeApp\ModelTyper\Actions;
 use FumeApp\ModelTyper\Traits\ClassBaseName;
 use FumeApp\ModelTyper\Traits\ModelRefClass;
 use Illuminate\Support\Collection;
+use ReflectionClass;
 use Symfony\Component\Finder\SplFileInfo;
 
 class GenerateJsonOutput
 {
+    /**
+     * @var array<string, array<string, mixed>>
+     */
     protected array $output = [];
 
+    /**
+     * @var array<int, ReflectionClass>
+     */
     protected array $enumReflectors = [];
 
     use ClassBaseName;
@@ -19,17 +26,24 @@ class GenerateJsonOutput
     /**
      * Output the command in the CLI as JSON.
      *
-     * @param  Collection<int, SplFileInfo>  $models
-     * @return string
+     * @param  Collection<int, \Symfony\Component\Finder\SplFileInfo>  $models
+     * @param  array<string, string>  $mappings
      */
-    public function __invoke(Collection $models): string
+    public function __invoke(Collection $models, array $mappings, bool $useEnums = false): string
     {
         $modelBuilder = app(BuildModelDetails::class);
         $colAttrWriter = app(WriteColumnAttribute::class);
         $relationWriter = app(WriteRelationship::class);
         $enumWriter = app(WriteEnumConst::class);
 
-        $models->each(function (SplFileInfo $model) use ($modelBuilder, $colAttrWriter, $relationWriter) {
+        $models->each(function (SplFileInfo $model) use ($modelBuilder, $colAttrWriter, $relationWriter, $mappings, $useEnums) {
+            $modelDetails = $modelBuilder($model);
+
+            if ($modelDetails === null) {
+                // skip iteration if model details could not be resolved
+                return;
+            }
+
             [
                 'reflectionModel' => $reflectionModel,
                 'name' => $name,
@@ -37,19 +51,19 @@ class GenerateJsonOutput
                 'nonColumns' => $nonColumns,
                 'relations' => $relations,
                 'interfaces' => $interfaces,
-            ] = $modelBuilder($model);
+            ] = $modelDetails;
 
             $this->output['interfaces'][$name] = $columns
                 ->merge($nonColumns)
                 ->merge($interfaces)
-                ->map(function ($att) use ($reflectionModel, $colAttrWriter) {
-                    [$property, $enum] = $colAttrWriter(reflectionModel: $reflectionModel, attribute: $att, jsonOutput: true);
+                ->map(function ($att) use ($reflectionModel, $colAttrWriter, $mappings, $useEnums) {
+                    [$property, $enum] = $colAttrWriter(reflectionModel: $reflectionModel, mappings: $mappings, attribute: $att, jsonOutput: true, useEnums: $useEnums);
                     if ($enum) {
                         $this->enumReflectors[] = $enum;
                     }
 
                     return $property;
-                });
+                })->toArray();
 
             $this->output['relations'] = $relations->map(function ($rel) use ($relationWriter, $name) {
                 $relation = $relationWriter(relation: $rel, jsonOutput: true);
@@ -60,11 +74,11 @@ class GenerateJsonOutput
                         'type' => 'export type ' . $relation['type'] . ' = ' . 'Array<' . $name . '>',
                     ],
                 ];
-            });
+            })->toArray();
         });
 
-        $this->output['enums'] = collect($this->enumReflectors)->map(function ($enum) use ($enumWriter) {
-            $enumConst = $enumWriter(reflection: $enum, jsonOutput: true);
+        $this->output['enums'] = collect($this->enumReflectors)->map(function ($enum) use ($enumWriter, $useEnums) {
+            $enumConst = $enumWriter(reflection: $enum, jsonOutput: true, useEnums: $useEnums);
 
             return [
                 $enumConst['name'] => [
@@ -72,8 +86,8 @@ class GenerateJsonOutput
                     'type' => $enumConst['type'],
                 ],
             ];
-        });
+        })->toArray();
 
-        return json_encode($this->output) . PHP_EOL;
+        return json_encode($this->output, \JSON_PRETTY_PRINT) . PHP_EOL;
     }
 }
