@@ -2,10 +2,12 @@
 
 namespace FumeApp\ModelTyper\Actions;
 
+use FumeApp\ModelTyper\Exceptions\ModelTyperException;
 use Illuminate\Support\Collection;
-use Illuminate\Support\Facades\File;
-use Symfony\Component\Finder\SplFileInfo;
 
+/**
+ * @throws \FumeApp\ModelTyper\Exceptions\ModelTyperException
+ */
 class Generator
 {
     /**
@@ -13,44 +15,57 @@ class Generator
      *
      * @return string
      */
-    public function __invoke(?string $specificModel = null, bool $global = false, bool $json = false, bool $plurals = false, bool $apiResources = false, bool $optionalRelations = false, bool $noRelations = false, bool $noHidden = false, bool $timestampsDate = false, bool $optionalNullables = false)
+    public function __invoke(?string $specificModel = null, bool $global = false, bool $json = false, bool $useEnums = false, bool $plurals = false, bool $apiResources = false, bool $optionalRelations = false, bool $noRelations = false, bool $noHidden = false, bool $timestampsDate = false, bool $optionalNullables = false, bool $fillables = false, string $fillableSuffix = 'Fillable')
     {
-        $models = $this->getModels($specificModel);
+        $models = app(GetModels::class)($specificModel);
 
-        return $this->display($models, $global, $json, $plurals, $apiResources, $optionalRelations, $noRelations, $noHidden, $timestampsDate, $optionalNullables);
-    }
-
-    /**
-     * Return collection of models.
-     *
-     * @return Collection<int, SplFileInfo>
-     */
-    protected function getModels(?string $model = null): Collection
-    {
-        return collect(File::allFiles(app_path('Models')))
-            ->filter(fn (SplFileInfo $file) => $file->getExtension() === 'php')
-            ->when(
-                $model,
-                fn ($files, $model) => $files->filter(fn (SplFileInfo $file) => $file->getBasename('.php') === $model)
-            )
-            ->values();
-    }
-
-    /**
-     * Display the command output.
-     *
-     * @param  Collection<int, SplFileInfo>  $models
-     */
-    protected function display(Collection $models, bool $global = false, bool $json = false, bool $plurals = false, bool $apiResources = false, bool $optionalRelations = false, bool $noRelations = false, bool $noHidden = false, bool $timestampsDate = false, bool $optionalNullables = false): string
-    {
         if ($models->isEmpty()) {
-            return 'ERROR: No models found.';
+            throw new ModelTyperException('No models found.');
         }
+
+        return $this->display(
+            models: $models,
+            global: $global,
+            json: $json,
+            plurals: $plurals,
+            apiResources: $apiResources,
+            optionalRelations: $optionalRelations,
+            noRelations: $noRelations,
+            noHidden: $noHidden,
+            timestampsDate: $timestampsDate,
+            optionalNullables: $optionalNullables,
+            useEnums: $useEnums,
+            fillables: $fillables,
+            fillableSuffix: $fillableSuffix
+        );
+    }
+
+    /**
+     * Return the command output.
+     *
+     * @param  Collection<int, \Symfony\Component\Finder\SplFileInfo>  $models
+     */
+    protected function display(Collection $models, bool $global = false, bool $json = false, bool $useEnums = false, bool $plurals = false, bool $apiResources = false, bool $optionalRelations = false, bool $noRelations = false, bool $noHidden = false, bool $timestampsDate = false, bool $optionalNullables = false, bool $fillables = false, string $fillableSuffix = 'Fillable'): string
+    {
+        $mappings = app(GetMappings::class)(setTimestampsToDate: $timestampsDate);
 
         if ($json) {
-            return app(GenerateJsonOutput::class)($models);
+            return app(GenerateJsonOutput::class)(models: $models, mappings: $mappings, useEnums: $useEnums);
         }
 
-        return app(GenerateCliOutput::class)($models, $global, $plurals, $apiResources, $optionalRelations, $noRelations, $noHidden, $timestampsDate, $optionalNullables);
+        return app(GenerateCliOutput::class)(
+            models: $models,
+            mappings: $mappings,
+            global: $global,
+            useEnums: $useEnums,
+            plurals: $plurals,
+            apiResources: $apiResources,
+            optionalRelations: $optionalRelations,
+            noRelations: $noRelations,
+            noHidden: $noHidden,
+            optionalNullables: $optionalNullables,
+            fillables: $fillables,
+            fillableSuffix: $fillableSuffix
+        );
     }
 }
