@@ -9,12 +9,9 @@ class WriteEnumConst
     /**
      * Write the enum const to the output.
      *
-     * @param  ReflectionClass  $reflection
-     * @param  string  $indent
-     * @param  bool  $jsonOutput
-     * @return array|string
+     * @return array{type: string, name: string}|string
      */
-    public function __invoke(ReflectionClass $reflection, string $indent = '', bool $jsonOutput = false): array|string
+    public function __invoke(ReflectionClass $reflection, string $indent = '', bool $jsonOutput = false, bool $useEnums = false): array|string
     {
         $entry = '';
 
@@ -31,11 +28,15 @@ class WriteEnumConst
         $cases = collect($reflection->getConstants());
 
         if ($cases->isNotEmpty()) {
-            $entry .= "{$indent}const {$reflection->getShortName()} = {\n";
+            if ($useEnums) {
+                $entry .= "{$indent}export const enum {$reflection->getShortName()} {" . PHP_EOL;
+            } else {
+                $entry .= "{$indent}const {$reflection->getShortName()} = {" . PHP_EOL;
+            }
 
-            $cases->each(function ($case) use ($indent, &$entry, $comments) {
+            $cases->each(function ($case) use ($indent, &$entry, $comments, $useEnums) {
                 $name = $case->name;
-                $value = is_string($case->value) ? "'{$case->value}'" : $case->value;
+                $value = is_string($case->value) ? "'" . addslashes($case->value) . "'" : $case->value;
 
                 // write comments if they exist
                 if (! empty($comments)) {
@@ -44,17 +45,26 @@ class WriteEnumConst
                             $comment = str_replace($name, '', $comment);
                             $comment = preg_replace('/[^a-zA-Z0-9\s]/', '', $comment);
                             $comment = trim($comment);
-                            $entry .= "{$indent}  /** $comment */\n";
+                            $entry .= "{$indent}  /** $comment */" . PHP_EOL;
                             break;
                         }
                     }
                 }
 
-                $entry .= "{$indent}  {$name}: {$value},\n";
+                if ($useEnums) {
+                    $entry .= "{$indent}  {$name} = {$value}," . PHP_EOL;
+                } else {
+                    $entry .= "{$indent}  {$name}: {$value}," . PHP_EOL;
+                }
             });
 
-            $entry .= "{$indent}} as const;\n\n";
-            $entry .= "{$indent}export type {$reflection->getShortName()} = typeof {$reflection->getShortName()}[keyof typeof {$reflection->getShortName()}]\n\n";
+            if ($useEnums) {
+                $entry .= "{$indent}}" . PHP_EOL . PHP_EOL;
+                $entry .= "{$indent}export type {$reflection->getShortName()}Enum = `\${{$reflection->getShortName()}}`" . PHP_EOL . PHP_EOL;
+            } else {
+                $entry .= "{$indent}} as const;" . PHP_EOL . PHP_EOL;
+                $entry .= "{$indent}export type {$reflection->getShortName()} = typeof {$reflection->getShortName()}[keyof typeof {$reflection->getShortName()}]" . PHP_EOL . PHP_EOL;
+            }
         }
 
         if ($jsonOutput) {
