@@ -3,8 +3,13 @@
 namespace FumeApp\ModelTyper\Commands;
 
 use FumeApp\ModelTyper\Actions\Generator;
+use FumeApp\ModelTyper\Exceptions\ModelTyperException;
 use Illuminate\Console\Command;
+use Illuminate\Filesystem\Filesystem;
+use Illuminate\Support\Facades\Config;
+use Symfony\Component\Console\Attribute\AsCommand;
 
+#[AsCommand(name: 'model:typer')]
 class ModelTyperCommand extends Command
 {
     /**
@@ -20,9 +25,11 @@ class ModelTyperCommand extends Command
      * @var string
      */
     protected $signature = 'model:typer
-                            {--model= : Generate your interfaces for a specific model}
-                            {--global : Generate your interfaces in a global namespace named models}
+                            {output-file? : Echo the definitions into a file}
+                            {--model= : Generate typescript interfaces for a specific model}
+                            {--global : Generate typescript interfaces in a global namespace named models}
                             {--json : Output the result as json}
+                            {--use-enums : Use typescript enums instead of object literals}
                             {--plurals : Output model plurals}
                             {--no-relations : Do not include relations}
                             {--optional-relations : Make relations optional fields on the model type}
@@ -30,21 +37,21 @@ class ModelTyperCommand extends Command
                             {--timestamps-date : Output timestamps as a Date object type}
                             {--optional-nullables : Output nullable attributes as optional fields}
                             {--api-resources : Output api.MetApi interfaces}
-                            {--all : Enable all output options (equivalent to --plurals --api-resources)}';
+                            {--fillables : Output model fillables}
+                            {--fillable-suffix= : Appends to fillables}
+                            {--ignore-config : Ignore options set in config}';
 
     /**
      * The console command description.
      *
      * @var string
      */
-    protected $description = 'Generate interfaces for all found models';
+    protected $description = 'Generate typescript interfaces for all found models';
 
     /**
      * Create a new command instance.
-     *
-     * @return void
      */
-    public function __construct()
+    public function __construct(protected Filesystem $files)
     {
         parent::__construct();
     }
@@ -54,20 +61,55 @@ class ModelTyperCommand extends Command
      */
     public function handle(Generator $generator): int
     {
-        // determine Laravel version
-        $laravelVersion = (float) app()->version();
+        try {
+            $output = $generator(
+                specificModel: $this->option('model'),
+                global: $this->getConfig('global'),
+                json: $this->getConfig('json'),
+                useEnums: $this->getConfig('use-enums'),
+                plurals: $this->getConfig('plurals'),
+                apiResources: $this->getConfig('api-resources'),
+                optionalRelations: $this->getConfig('optional-relations'),
+                noRelations: $this->getConfig('no-relations'),
+                noHidden: $this->getConfig('no-hidden'),
+                timestampsDate: $this->getConfig('timestamps-date'),
+                optionalNullables: $this->getConfig('optional-nullables'),
+                fillables: $this->getConfig('fillables'),
+                fillableSuffix: $this->getConfig('fillable-suffix'),
+            );
 
-        if ($laravelVersion < 9.20) {
-            $this->error('This package requires Laravel 9.20 or higher.');
+            /** @var string|null $path */
+            $path = $this->argument('output-file');
+
+            if (is_null($path) && Config::get('modeltyper.output-file', false)) {
+                $path = (string) Config::get('modeltyper.output-file-path', '');
+            }
+
+            if (! is_null($path) && strlen($path) > 0) {
+                $this->files->ensureDirectoryExists(dirname($path));
+                $this->files->put($path, $output);
+
+                $this->info('Typescript interfaces generated in ' . $path . ' file');
+
+                return Command::SUCCESS;
+            }
+
+            $this->line($output);
+        } catch (ModelTyperException $exception) {
+            $this->error($exception->getMessage());
 
             return Command::FAILURE;
         }
 
-        $plurals = $this->option('plurals') || $this->option('all');
-        $apiResources = $this->option('api-resources') || $this->option('all');
-
-        echo $generator($this->option('model'), $this->option('global'), $this->option('json'), $plurals, $apiResources, $this->option('optional-relations'), $this->option('no-relations'), $this->option('no-hidden'), $this->option('timestamps-date'), $this->option('optional-nullables'));
-
         return Command::SUCCESS;
+    }
+
+    private function getConfig(string $key): string|bool
+    {
+        if ($this->option('ignore-config')) {
+            return $this->option($key);
+        }
+
+        return $this->option($key) ?: Config::get("modeltyper.{$key}");
     }
 }
