@@ -9,13 +9,11 @@ trait ModelRefClass
     /**
      * Get the reflection interface.
      *
-     * @param  array  $info - The model details from the model:show command.
-     * @return ReflectionClass
+     * @param  array{"class": class-string<\Illuminate\Database\Eloquent\Model>, database: string, table: string, policy: class-string|null, attributes: \Illuminate\Support\Collection, relations: \Illuminate\Support\Collection, events: \Illuminate\Support\Collection, observers: \Illuminate\Support\Collection, collection: class-string<\Illuminate\Database\Eloquent\Collection<\Illuminate\Database\Eloquent\Model>>, builder: class-string<\Illuminate\Database\Eloquent\Builder<\Illuminate\Database\Eloquent\Model>>}  $info
+     * @return \ReflectionClass<\Illuminate\Database\Eloquent\Model>
      */
     public function getRefInterface(array $info): ReflectionClass
     {
-        $class = $info['class'];
-
-        return new ReflectionClass($class);
+        return new ReflectionClass($info['class']);
     }
 }
